requirements.txt
pillow
tensorflow
numpy
flask
